// shared/types/server/managerDto.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Omnition Api Documentation
 * Документация Api для Omnition
 * OpenAPI spec version: 1.0
 */

export interface ManagerDto {
    /** ID менеджера */
    id: number;
    /** Имя менеджера */
    name: string;
    /** Должность менеджера */
    post: string;
    /** Телефон менеджера */
    phone: string;
    /** Email менеджера */
    email: string;
    /** Фото менеджера */
    photo: string;
}